(REF) 가게 관리 화면(Store Management) MVI 구조 및 도메인 로직 분리 (#20)
